Move the worker entry point to the root folder
