+++ api/package.json
@@ -150,7 +150,7 @@
     "scalingo-postbuild": "node scripts/generate-cron > cron.json",
     "start": "node index.js",
     "start:watch": "nodemon index.js",
-    "start:job": "node lib/application/run-jobs.js",
+    "start:job": "node worker.js",
     "test": "NODE_ENV=test npm run db:prepare && npm run test:api",
     "test:api": "status=0; npm run test:api:unit || status=1 ; for dir in $(find tests/* -maxdepth 0 -type d -not -path tests/unit) ; do npm run test:api:path -- $dir || status=1 ; done ; exit $status",
     "test:api:path": "NODE_ENV=test mocha --exit --recursive --reporter=${MOCHA_REPORTER:-dot}",
